gradesApp/requirements.txt
python-dotenv
pymysql
peewee
flask
Flask-Session
redis
